Update: small details and navigator screens
