+++ lib/ui/screens/login_screen.dart
@@ -171,7 +171,9 @@ class _LoginScreenState extends State<LoginScreen> {
           fontSize: 30,
         ),
       ),
-      onPressed: () {},
+      onPressed: () {
+        Navigator.popAndPushNamed(context, '/home');
+      },
       icon: const Icon(
         Icons.check,
         size: 40,
@@ -183,7 +185,7 @@ class _LoginScreenState extends State<LoginScreen> {
   _buildLink() {
     return GestureDetector(
       onTap: () {
-        debugPrint('Foi clicado');
+        Navigator.popAndPushNamed(context, '/register');
       },
       child: Container(
         padding: const EdgeInsets.only(
@@ -195,7 +197,7 @@ class _LoginScreenState extends State<LoginScreen> {
           style: TextStyle(
             fontSize: 15,
             decoration: TextDecoration.underline,
-            //color: AppColors.lilas,
+            color: AppColors.lilas,
           ),
         ),
       ),
