+++ lib/ui/widgets/app_drawer.dart
@@ -17,7 +17,9 @@ class AppDrawer extends StatelessWidget {
               child: _buildListTile(
                 icon: Icons.account_circle_outlined,
                 text: 'Perfil',
-                onTap: () {},
+                onTap: () {
+                  Navigator.popAndPushNamed(context, '/profile');
+                },
               ),
             ),
             Container(
